- 基本完成了首页的设计实现 - todo 动态加载 / 特殊情况处理，如无卡，一张卡 / 未登录状态处理 / Relative布局

+++ APP/Android/PointsLeague/app/src/main/java/com/citiexchangeplatform/pointsleague/models/HistoryExchangeItemModel.java
@@ -2,10 +2,10 @@ package com.citiexchangeplatform.pointsleague.models;
 
 public class HistoryExchangeItemModel {
     private int point;
-    private int exchangePoint;
+    private double exchangePoint;
     private String date;
 
-    public HistoryExchangeItemModel(int point, int exchangePoint, String date) {
+    public HistoryExchangeItemModel(int point, double exchangePoint, String date) {
         this.point = point;
         this.exchangePoint = exchangePoint;
         this.date = date;
@@ -15,7 +15,7 @@ public class HistoryExchangeItemModel {
         return point;
     }
 
-    public int getExchangePoint() {
+    public double getExchangePoint() {
         return exchangePoint;
     }
